requirements.txt
opencv-python==4.8.0
PyQt5==5.15.9
numpy==1.24.3
pillow==10.0.0